+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Semantic Versioning](http://semver.org/).
 
 ## 1.3.9.dev (development stage/unreleased)
+### Added
+- exception for `asyncio.base_futures.InvalidStateError` https://github.com/unicorn-data-analysis/unicorn-binance-websocket-api/issues/18
+### Changed
+- create_stream() returns False if websocket URL is to long
+
 
 ## 1.3.9
 ### Changed